backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "FastAPI backend service"
authors = [
    {name = "Developer", email = "developer@example.com"},
]
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.3.5",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.4",
    "PyPDF2>=3.0.0",
    "tiktoken>=0.5.0",
    "numpy>=1.21.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
    "pyodbc>=4.0.35",
    "sqlalchemy>=2.0.0",
    "aiohttp>=3.8.0",
    "sentence-transformers>=2.2.0",
]
readme = "README.md"
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]
